requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.3
openai==1.12.0
httpx==0.24.1
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0